askyourdata/requirements.txt
streamlit
torch
transformers
pandas
openai
transformers
torch
sentencepiece
llama-cpp-python
xlsxwriter